requirements.txt
azure-identity==1.14.0
Flask==2.3.2
openai==0.28.0
azure-search-documents==11.4.0b6
azure-storage-blob==12.17.0
python-dotenv==1.0.0
azure-cosmos==4.5.0
SQLAlchemy==1.4.29
psycopg2==2.9.7
requests==2.31.0
Flask-Executor==1.0.0
tiktoken==0.4.0
PyPDF2~=3.0.1
pdfminer.six==20221105
